Fix some mistranslations
Among other things, EP03_S0040 has had quite a few lines cut (on
2021-02-22 according to comments in the Lua script), including
EP03_S0040_0068 and EP03_S0040_0076, but the English script was
never updated to flow well without them. Edit those lines to make
it work now but *also* to make it work if by some miracle the cut
lines should be reintroduced in a future patch or mod.