Tweak a line to restore original ambiguity (and improve punctuation)
It's not clear from the Japanese written line and voice line
whether he lists three alternatives:

  1. you came to contact us
  2. you came for the library
  3. you came for Miyuki's sake

or two alternatives with an added elaborating phrase:

  1. you came to contact us
  2. you came for the library... which is perhaps
     somehow necessary for Miyuki's sake?

Crosschecking with the Eastern translations, the Chinese translation
seems to say the former explicitly, whereas the Korean version seems
to match the Japanese exactly. Of course, the Western translations,
as seen here, chose the latter option. So it looks like there is no
consensus and it really is as ambiguous as it seems to me.