Edit a line in hopes of letting it make more sense
This line is odd in all languages. But I think "holding herself aloof"
is the opposite of desiring cooperation, so make it more ambiguous. At
the same time, intensify the first clause to make it clearer that Aoi
felt there was so little cooperation that she had to insist on it,
which I think is implied here: more plainly, the Japanese seems to say
"Aoi had a penchant for stubbornly demanding that we cooperate".

I suspect one of these two meanings was intended originally:

* "Aoi urged cooperation but you didn't comply, so you might think
   she'd end up distancing herself from you in anger, but she didn't."

* "You might think Aoi urged cooperation because she felt you weren't
   close enough to her, but I don't think she felt like that."

  This is somewhat supported by the use of "wake de wa nai"
  in the second clause, but this seems to fit the verb worse.

The Japanese verb used here is defined in a Japanese dictionary as
"avoid someone you dislike; keep away and avoid getting intimate".