Remove the game's only mention of "Baby"/"In-Training"
This evolution stage is not available in-game (despite being named in
SYS_WORD_4054), so this term is never explained and only ever gets
mentioned in this profile. In an attempt to be the most readable and
accessible to someone whose first entry in the franchise might be Survive,
edit the profile by replacing this term with a more common word.

Reproduced-by: Oleg Oshmyan <chortos@inbox.lv>
Reproduced-on: Sat Oct 7 21:35:46 2023 +0300