Species profiles: avoid the word "data" to make sense in-universe
This mention of "data" is present in the Japanese version as well,
but the official profiles have already changed all references
to Digimon to "monsters" and Digital World to "the other world",
so this one sticks out like a sore thumb.

The fuller phrase in the Japanese is: フィールド中の鉱石データをまとい.
Besides "data", it also mentions a "field", which the official translation
already omits. This is likely a reference to https://wikimon.net/Field.
The profile is old and dates back to when Fields were a thing in Digimon
(the PS2 Savers game, which came out the year before Gottsumon got added
to the reference book, makes use of them).

Reproduced-by: Oleg Oshmyan <chortos@inbox.lv>
Reproduced-on: Sat Oct 7 22:05:38 2023 +0300