Rewrite a previously-fixed line
This is the very moment Takuma realizes there's more
than one creature, so the focus should be on that.
The new wording loses any explicitness of "before we knew it",
but I think it is implied enough by the surprised expression.

Besides, is "corner" really an appropriate verb here?
"Surround" would have been better.