Change a sentence subject for better dialogue flow
The Japanese text does not specify the subject here. Technically,
both "you" and "he" work, depending on who Takuma is addressing
in this sentence: Kaito or everyone else. But he clearly addresses
Kaito in the second sentence on the very same line, so it's most
natural to assume that he does so throughout the entire line.