Fix mistranslated/altered Dokugumon exposition in late Part 1
For the most part of this dialogue, the Boy refers to "them" in plural
in Japanese, and indeed, we see multiple Dokugumon later, so use plurals
in English, too. However, in several lines, he uses the singular form.

It's not entirely clear if it is intended to contrast with the plural, as
Japanese usually doesn't distinguish between singular and plural at all,
so while this is one of the exceptions that do have singular and plural
forms, there's a chance that the singular "guy" here is just intended as
a generic placeholder, similar to "the kind of guy" or implicitly standing
for every one of "them".

However, there's also a chance that the singular form is meant to refer
to the particular Dokugumon individual that took Aoi. This seems a little
odd, but if this wasn't intended, then the switching between singular and
plural itself seems a little odd as well. This is further corroborated by
Takuma acting surprised during the subsequent battle that there is more
than one Dokugumon: it's still weird, given that the Boy clearly speaks of
plural "them", but the Boy never explicitly says that "they" are multiple
*spiders*, rather than, for example, various kinds of creatures that
include a single spider.

I've done my best to maintain/restore the distinction in English, trying
to preserve as much of the ambiguity as possible but ultimately assuming
that the singular form refers to the specific Dokugumon that took Aoi.

-

Reword EP01_S0800_0046 to avoid the interpretation that the whole *school*
is full of webs: we know that it is, but the Boy is specifically talking
about the gym and about how the webs are blocking the gym off.

-

I doubt that the crawl space entrance can be "secret". In this case,
抜け穴 is used more in its literal sense: it's a hole that you drop into.
In the next scene, it does indeed appear hidden, as Takuma seems to break
some boards of the flooring to unveil the hole, but perhaps those are just
loose boards that can be put back? Hidden is not the same as secret, in
any case. And if it were not resealable, why would the Boy know about it?
And why wouldn't he recommend a different entrance that was already open?
All in all, I think the Japanese text just uses the simplest word that
applies, so just use the simplest natural wording in English, too.