Fix some mistranslations in Part 1
EP01_S0380_0203 and EP01_S0380_0204 are unused copies of EP01_S0380_0202.

EP01_S0801_0025 and EP01_S0801_0026 are the same regardless of how many
people have been captured, and their Japanese version doesn't specify
whether they're talking about Aoi alone or (potentially) all the captives.
It's impossible to make 0025 less explicit in English, so stick to "Aoi",
which thankfully makes *some* sense because Aoi has been abducted
for the longest amount of time and hence her health is at most risk.
However, 0026 can be reworded to avoid specifying the object
and to apply equally well to Aoi and to everyone, so do that.