Further correct & improve previously edited lines in early Part 2
In Part 1, Aoi may have been captured alone, or additional people and
Digimon may have been captured later (depending on the player's actions).
Rephrase this line to work in all cases.

Also just make the dialogue flow more naturally.
Capitalize on the fact that in the audio (i. e. in the original Japanese),
Haru is mentioned only at the end of the first line, which prompts Agumon
to remember about him suddenly (with an accompanying visual effect)
and to speak the second line: move his mention similarly to the end
of the English line.