Improve a few lines in Part 5
Stick to a single term for 幻: Takuma is repeating what he's just heard,
so it's weird for him to change "illusions" to "visions".

These two lines and the line immediately following them are displayed
if *at least one* item was discovered, but the game's script makes no
distinction between the singular and plural cases, so try to word these
lines in such a way that they work regardless of how many items were
actually discovered. I think I have achieved this for these two lines,
but the third line seems impossible to rephrase in this manner. This is
very unfortunate, but we can find some small solace in the fact that this
issue equally applies to the Japanese text here: its third line (and only
its third line) speaks explicitly of plural people and plural items.

While we're at it, add "just" to the first line to make it closer
to the original text.