Revert "Fix bad plural in "Episodes 1" etc. in character profile UI"
This reverts commit a9812dc82b1a21ffc1f597320407f5fce0fec2fb.

Turns out the same string is shown in the tab's heading (without any
numeric suffix). I guess it's better to show an "Episodes" overall
heading with "Episodes 1/2/3" in the list (which with some imagination
can be interpreted as "Episodes, part 1/2/3") than "Episode 1/2/3" with
singular "Episode" in the overall heading.

Perhaps this should be renamed to some uncountable noun.