Fix mistranslations in post-battle Gomamon dialogue in Part 2
Note: this dialogue is branched, depending on how many Gomamon are alive.

Gomamon C mentions the 祠, which is consistently translated as
"inner shrine" since commit 79aadf9668236c9457f87dcc8a1b52efa2a65436.
However, it would seem odd for a Gomamon to know this term,
so use "small shrine" as an exception.