Consistently render the overground 水路 a "canal"
Most of the current dialogue lines in Part 2 call it a "canal",
so stick to this term to avoid confusion.

Not "aqueduct", because it's neither elevated nor
apparently intended for irrigation or drinking water supply.

Not "waterway", because only one existing line calls it that.
This also helps differentiate it from the "underground waterway",
although the Japanese version is inconsistent about this: the overground
one is always 水路, but the underground one is both 水道 and 水路.