Clarify "Dam - Underground" map area label
The full Japanese label says: "Dam/Underground Waterway Area".

Separating them with a dash makes it look as though this is some
underground part of the dam, but no such thing even exists.

"Dam/Underground Waterway" is too long to fit (on one line) in the map
heading box or in the choice box in the area selection list (at least when
additional icons are shown that signify Free Battle and story progress).

"Dam/Waterway Area" fits nicely, reads nicely, and captures the gist
of the Japanese label, even though the "underground" part is dropped.