Rephrase Labramon's explanation of 'deviant' to make more sense
In Japanese, this is a straightforward pun: 変態 means both
"metamorphosis" and "pervert". In English, I haven't been able to find
an equivalent, and neither have the official translators, clearly.
After some consideration, my best idea is to keep the official choice
of "deviant" but to edit Labramon's lines to stray slightly farther from
the Japanese but to make more sense in the context of the English word.

Also, it seems weird for Floramon to praise "you both" when Kunemon just
kewws and at most agrees with what Labramon said. It can also be
misinterpreted as "Labramon and Agumon", which wouldn't be right.
The Japanese doesn't specify "both", so drop it here.