Reword a line that mentions sacrifices in early Part 3
This seems to be an error in the Japanese script: Fangmon never says
anything about sacrifices and is never shown to be aligned with the group
that does seek child sacrifices. The only sign that he might want them is
that his Gazimon henchmen left the Professor to go after more kids (us).

That said, this line is a little milder in Japanese than it is in the
official English, so I'll use this as an excuse to make the English line
milder, too. But in order to stay faithful to the audio and the Japanese
script, I'm not fully removing the mention. I'm hoping that this can be
interpreted as Minoru's impression/recollection of Fangmon's threats
to drop the kids off the dam or of some additional remark that was left
off screen in Part 2 proper.

While I'm at it, make this line even *more* faithful to the Japanese
script by preserving "turning us human kids into sacrifices" verbatim.