Reword a line to let Takuma make less of a fool of himself
You asked her insistently. Why are you acting surprised when she answered?
Did you actually expect her to ignore you? If you did, why did you insist
so much on asking in the first place, huh?!

I think Takuma's mainly surprised at how *well* she spoke, not that she
answered at all. The Japanese line is like, "She talked just now?!" which
is also weird but can be interpreted, with some benefit of the doubt,
as Takuma expecting a nonverbal response, as is common for Miyuki;
but that's still a response, an answer.