Fix mistranslations in Part 5 Free Action
EP05_S1400_0016 may be accidentally copied from EP05_S1300_0027,
but the dialogue is weird and inaccurate beyond this one line anyway.

'What's a "rocker"?' is cool, but it's not clear why that should be deep.
Defining "common sense" (as the Japanese text has), on the other hand,
is indeed a deep topic, so stick to a literal translation there.
It works nicely with my new, more neutral translation of EP05_S1400_0015,
and "common sense" is odd enough of a phrase in English
to make Agumon asking about it quite plausible.