Fix some more mistranslations in Part 5 Waterway
Some of these *may* have been correct before late script changes.
Still, they aren't correct now: the choice of path doesn't matter;
what matters is the choice of Digimon; and EP05_S1941 is available
as soon as we've met the first fake Labramon, not only after the choice,
but it does double duty because it's also shoved in the player's face
immediately after the choice if they do take the wrong pick.
So it has to work well both if Aoi has seen only one (fake) Labramon
and if she's seen both and chosen poorly.