Tweak a line in Part 5 Waterway to avoid unintended implication
If this dialogue has been seen before reaching the real Floramon,
but then the player picks the wrong Floramon, they come back to
this location and Saki laments her failure to detect the real one.
If the player activates this dialogue immediately after, then the
"Whatever" makes it sound like Saki has stopped caring about Floramon,
which is completely wrong.