Fix a few previously-neglected lines in Wrathful Part 11
Not sure how I missed it before, but EP11_S7031_0035 used 'she' for
Renamon, so this line has been reported by an attentive reader. Thanks!
But it's too weird to refer to Renamon in third person in this line
in the first place, because it's literally being said *to* Renamon.

EP11_S7031_0009 was way too long for the corresponding audio and had
a whole sentence that's entirely absent from the Japanese version.
EP11_S7031_0008 also had more words and explicitness than the Japanese
in order to link better with the official 0009, so edit them both
to bring them back in line with the Japanese text and audio.